[RAPPS-DB] Downgrade Hiawatha to 9.15 to fix CORE-13937
Starting from Hiawatha 10.0 their downloads are 64bit and therefore
not suited for ReactOS.

The older 9.14 that we had until around 2016 is broken too,
because they used hard-coded paths in their Hiawatha.bat
that do not necessarily exist in localized Windows or
ReactOS versions.

Therefore sticking to 9.15 seems to be a good idea.