[RAPPS-DB] Downgrade PdfSam 3.3.5 to 2.2.4
v3.3.5 requires Java 1.8.0 and that does not run yet on ReactOS and never did up
to now. Until we have the prerequisites running, it makes no sense to have
PdfSam 3.3.5 as a non.working app in rapps.
PdfSam 2.2.4 can be started up with both from our rapps Java versions:
Java 6u45 & 7u80. So let's stick with that older version for now.