This is a slightly modified version of winfile so that it can be compiled by mingw for Windows Server 2003 SP2. There is a small hack that allows it to run on ReactOS.
GNUmakefile was modified so that it can cross compiled on linux for
Windows.
This release was compiled on Ubuntu 18.04 with

i686-w64-mingw32-g++ (GCC) 7.3-win32 20180312
i686-w64-mingw32-gcc (GCC) 7.3-win32 20180312

The installer was created using NSIS v2.51 on Ubuntu 18.04 under wine.
The winfile.exe was statically linked so it's a stand alone binary.