[RAPPS-DB] DirectX June 2010 update URL
MS updated many binaries with SHA-2-signed ones
late 2020 or early 2021
and therefore we had hash mismatch.
Stick to the previous SHA-1-signed binary for now.
Allows to stick with same file-size also
and the downloads certificate can also still be
verified with XPSP3, which is not possible with
the new SHA-2-signed ones.

Ftr version tab shows 9.0.8112.16421.