[RAPPS-DB] Delete Disk Genius 5.1
The URL is down, for 1+ year already.
There is no backup of it
on web.archive.org and I couldn't quickly
find any alternative download location
for that Chinese tool.
It makes no sense to keep such corpses
within the database for years.
If you find a new download location,
feel free to re-add it.