[RAPPS-DB] LGeneral 1.4.4 & SDL
The old LGeneral 1.1 URL was down,
because the author moved it on sf.
The new package works as well in ros,
and has SDL-runtime & sdl-mixer
included already, which frees the user
to copy them into the games folder
retrospectively, and frees us from
the necessity to hold them in Rapps-DB
because we do have no other app that
would rely on them, without also
shipping them.
Only drawback of the new LGeneral 1.4.4
package is that it is zip and no exe anymore.