[RAPPS-DB] Delete LibreOffice 5.1.1.3
The newer 5.4.7.2, that we have in rapps
as well, works as 'good' as the older one
for all known JIRA tickets.
So no need to keep that dupe any longer.
Especially since testing both for each
release is very time consuming.

Even binary size comparison leaves Libre 5.1.1.3
just in the poor middle-field:

225.308.672bytes Libre 5.4.7.2
221.724.672bytes Libre 5.1.1.3
142.587.736bytes Apache OpenOffice 4.1.10

And none of the 3 works good enough for anything productive yet.