[RAPPS-DB] Remove Spotify 1.0.29.92 (#288)
Remove Spotify 1.0.29.92 from Rapps database because this offical client for Windows XP/Server 2003 is no longer supported by Spotify starting from December 2022. So now it has exactly zero functionality for us and hence is useless (except for regression testing probably). It does not allow to login with credentials and can't be used anymore.
I will replace it by SpotifyXP, which is an unofficial (3rd party) client written in Java, and it works in ReactOS, as it's made especially for XP/2003 and is the only way to listen Spotify music on these OSes.